feat: refactoring transportHandler in the internal thread
In order to have a 1 <-> 1 relationship between the IPC transport instantiated
in the renderer process and the transport instance living in the internal process.

The refector simplifies the transport logic in LLD by removing the usage of
`withDevice`/`withTransport` from the internal process.
For ex: if an error occurs while trying to open an HID USB transport, the error is
propagated back to the renderer thread without retrying, and it is inside the
renderer thread that it is decided to retry (or not) to open the transport.

to squash f0de7f36cdabd9825976802002540a3e9086403e